Refactor test code for global store/ktables examples
